refactor: update layout imports to use profile-layout for settings pages
