+++ resources/js/pages/settings/appearance.tsx
@@ -3,7 +3,7 @@ import HeadingSmall from '@/components/heading-small';
 import { type BreadcrumbItem } from '@/types';
 
 import AppLayout from '@/layouts/app-layout';
-import SettingsLayout from '@/layouts/settings/layout';
+import SettingsLayout from '@/layouts/settings/profile-layout';
 
 const breadcrumbs: BreadcrumbItem[] = [
   {
