+++ resources/js/pages/settings/profile.tsx
@@ -1,23 +1,14 @@
 import { type BreadcrumbItem, type SharedData } from '@/types';
 import { Transition } from '@headlessui/react';
 import { Link, useForm, usePage } from '@inertiajs/react';
-import { type FormEventHandler } from 'react';
+import { useMemo, type FormEventHandler } from 'react';
 
-import DeleteUser from '@/components/delete-user';
+import { InputField } from '@/components/forms/inputs/InputField';
 import HeadingSmall from '@/components/heading-small';
-import InputError from '@/components/input-error';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
 import AppLayout from '@/layouts/app-layout';
-import SettingsLayout from '@/layouts/settings/layout';
-
-const breadcrumbs: BreadcrumbItem[] = [
-  {
-    title: 'Profile settings',
-    href: route('profile.edit'),
-  },
-];
+import SettingsLayout from '@/layouts/settings/profile-layout';
+import { useLaravelReactI18n } from 'laravel-react-i18n';
 
 type ProfileForm = {
   name: string;
@@ -25,6 +16,7 @@ type ProfileForm = {
 };
 
 export default function Profile({ mustVerifyEmail, status }: { mustVerifyEmail: boolean; status?: string }) {
+  const { t } = useLaravelReactI18n();
   const { auth } = usePage<SharedData>().props;
 
   const { data, setData, patch, errors, processing, recentlySuccessful } = useForm<Required<ProfileForm>>({
@@ -39,69 +31,64 @@ export default function Profile({ mustVerifyEmail, status }: { mustVerifyEmail: 
       preserveScroll: true,
     });
   };
-
+  const breadcrumbs: BreadcrumbItem[] = useMemo(
+    () => [
+      {
+        title: t('Profile Settings'),
+        href: route('profile.edit'),
+      },
+    ],
+    [t],
+  );
   return (
-    <AppLayout title="Profile Settings" breadcrumbs={breadcrumbs}>
+    <AppLayout title={t('Profile Settings')} breadcrumbs={breadcrumbs}>
       <SettingsLayout>
         <div className="space-y-6">
-          <HeadingSmall title="Profile information" description="Update your name and email address" />
+          <HeadingSmall title={t('Profile information')} description={t('Update your name and email address')} />
 
           <form onSubmit={submit} className="space-y-6">
-            <div className="grid gap-2">
-              <Label htmlFor="name">Name</Label>
+            <InputField
+              label={t('Name')}
+              value={data.name}
+              onChange={(value) => setData('name', value)}
+              required
+              autoComplete="name"
+              placeholder={t('Full name')}
+              error={errors.name}
+            />
 
-              <Input
-                id="name"
-                className="mt-1 block w-full"
-                value={data.name}
-                onChange={(e) => setData('name', e.target.value)}
-                required
-                autoComplete="name"
-                placeholder="Full name"
-              />
-
-              <InputError className="mt-2" message={errors.name} />
-            </div>
-
-            <div className="grid gap-2">
-              <Label htmlFor="email">Email address</Label>
-
-              <Input
-                id="email"
-                type="email"
-                className="mt-1 block w-full"
-                value={data.email}
-                onChange={(e) => setData('email', e.target.value)}
-                required
-                autoComplete="username"
-                placeholder="Email address"
-              />
-
-              <InputError className="mt-2" message={errors.email} />
-            </div>
+            <InputField
+              label={t('Email address')}
+              value={data.email}
+              onChange={(value) => setData('email', value)}
+              required
+              autoComplete="username"
+              placeholder={t('Email address')}
+              error={errors.email}
+            />
 
             {mustVerifyEmail && auth.user.emailVerifiedAt === null && (
               <div>
                 <p className="text-muted-foreground -mt-4 text-sm">
-                  Your email address is unverified.{' '}
+                  {t('Your email address is unverified.')}
                   <Link
                     href={route('verification.send')}
                     method="post"
                     as="button"
                     className="text-foreground underline decoration-neutral-300 underline-offset-4 transition-colors duration-300 ease-out hover:decoration-current! dark:decoration-neutral-500"
                   >
-                    Click here to resend the verification email.
+                    {t('Click here to resend the verification email.')}
                   </Link>
                 </p>
 
                 {status === 'verification-link-sent' && (
-                  <div className="mt-2 text-sm font-medium text-green-600">A new verification link has been sent to your email address.</div>
+                  <div className="mt-2 text-sm font-medium text-green-600">{t('A new verification link has been sent to your email address.')}</div>
                 )}
               </div>
             )}
 
             <div className="flex items-center gap-4">
-              <Button disabled={processing}>Save</Button>
+              <Button disabled={processing}>{t('Save')}</Button>
 
               <Transition
                 show={recentlySuccessful}
@@ -110,13 +97,11 @@ export default function Profile({ mustVerifyEmail, status }: { mustVerifyEmail: 
                 leave="transition ease-in-out"
                 leaveTo="opacity-0"
               >
-                <p className="text-sm text-neutral-600">Saved</p>
+                <p className="text-sm text-neutral-600">{t('Saved')}</p>
               </Transition>
             </div>
           </form>
         </div>
-
-        <DeleteUser />
       </SettingsLayout>
     </AppLayout>
   );
